move Ashby MCP server to new metatada pattern
